refactor: profile upload handling to use parseMultipart utility and remove multer
