--- blog-backend/src/utils/uploads.ts
@@ -1,3 +1,0 @@
-import multer from 'multer';
-
-export const upload = multer({ storage: multer.memoryStorage() });
